pocketmoney/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
scipy==1.5.4
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.26.2
